refactor: extract clipboard functionality to support containerized environments
- Move clipboard code to dedicated utils/clipboard.py module
- Remove WSL2-specific PowerShell clipboard handling
- Add container-aware clipboard support via environment variables
- Support file-based clipboard for containerized deployments
- Update documentation with new clipboard configuration options
